f3s/tracing-demo/docker/backend/requirements.txt
flask==3.0.0
opentelemetry-distro==0.49b0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-instrumentation-flask==0.49b0
